+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The changes are relative to the previous release, unless the baseline is specifi
   avifGainMapMetadataDouble structs.
 * Add avif(Un)SignedFraction structs and avifDoubleTo(Un)SignedFraction
   utility functions.
+* Add customEncodeImageFunc, customEncodeFinishFunc and customEncodeData fields
+  to the avifEncoder struct to override the AV1 codec.
 
 ## [1.1.1] - 2024-07-30